# cmd/app/docs/swagger.yaml
basePath: /
definitions:
...
  financial_system_internal_domain.ActionLog:
    properties:
      action_type:
        type: string
      created_at:
        type: string
      details:
        items:
          type: integer
        type: array
      id:
        type: integer
      is_undone:
        type: boolean
      user_id:
        type: integer
    type: object
...
  internal_transport_rest.authInputDTO:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления финансами. Клиент, менеджер и администратор.
  title: Financial System API
  version: "1.0"
paths:
  /admin/logs:
    get:
      description: Возвращает все записи action_logs в порядке убывания времени.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/financial_system_internal_domain.ActionLog'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Все логи действий
      tags:
      - admin
  /admin/logs/{id}/undo:
    post:
      parameters:
      - description: ID записи лога
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отменить действие
      tags:
      - admin
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: 'Вход по email и паролю. Возвращает JWT для заголовка Authorization:
        Bearer &lt;token&gt;. Аккаунт должен быть подтверждён (is_active).'
      parameters:
      - description: email, password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_rest.authInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Регистрация нового клиента. Требует подтверждения менеджером (is_active).
      parameters:
      - description: email, password (мин. 8 символов)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_rest.authInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация клиента
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"